requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
flask-paginate==2022.1.8
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
pymongo==4.1.1
Werkzeug==2.1.2
python-dateutil==2.8.2
